+++ orangetool/__main__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+"""Orangetool main."""
+
+from .orangetool_system import version
+
+if __name__ == "__main__":
+    version()