+++ orangetool/orangetool_ip.py
@@ -1,14 +1,13 @@
 # -*- coding: utf-8 -*-
 """Orangetool IP functions."""
 from .orangetool_system import restart as restart_func
+from .orangetool_params import IP_PATTERN, GLOBAL_IP_API_1, GENERAL_ERROR_MESSAGE
 import subprocess as sub
 import socket
 import os
 import requests
 import re
 import platform
-ip_pattern = r"(?:[0-9]{1,3}\.){3}[0-9]{1,3}"
-api_1 = "http://ipinfo.io/ip"
 
 
 def internet(host="8.8.8.8", port=53, timeout=3):
@@ -16,10 +15,10 @@ def internet(host="8.8.8.8", port=53, timeout=3):
     Check internet connections.
 
     :param  host: the host that check connection to
-    :param  port: port that check connection with
-    :param  timeout: times that check the connection
     :type host:str
+    :param  port: port that check connection with
     :type port:int
+    :param  timeout: times that check the connection
     :type timeout:int
     :return bool: True if Connection is Stable
     >>> internet() # if there is stable internet connection
@@ -36,19 +35,19 @@ def internet(host="8.8.8.8", port=53, timeout=3):
         return False
 
 
-def local_ip(DEBUG=False):
+def local_ip(debug=False):
     """
     Return local ip of computer in windows by socket module and in unix with hostname command in shell.
 
-    :param DEBUG:flag for using Debug Mode
-    :type DEBUG:bool
+    :param debug:flag for using debug Mode
+    :type debug:bool
     :return: local ip as string
     """
     try:
         ip = socket.gethostbyname(socket.gethostname())
         if ip != "127.0.0.1":
             return ip
-        elif platform.system() != "Windows":
+        if platform.system() != "Windows":
             command = sub.Popen(["hostname",
                                  "-I"],
                                 stdout=sub.PIPE,
@@ -58,47 +57,47 @@ def local_ip(DEBUG=False):
             response = list(command.communicate())
             if len(response[0]) > 0:
                 return str(response[0])[2:-4]
-            return "Error"
-        return "Error"
+            return GENERAL_ERROR_MESSAGE
+        return GENERAL_ERROR_MESSAGE
 
     except Exception as e:
-        if DEBUG:
+        if debug:
             print(str(e))
-        return "Error"
+        return GENERAL_ERROR_MESSAGE
 
 
-def global_ip(DEBUG=False):
+def global_ip(debug=False):
     """
     Return ip with by http://ipinfo.io/ip api.
 
-    :param DEBUG:flag for using Debug mode
-    :type DEBUG:bool
+    :param debug:flag for using debug mode
+    :type debug:bool
     :return: global ip as string
     """
     try:
         new_session = requests.session()
-        response = new_session.get(api_1)
-        ip_list = re.findall(ip_pattern, response.text)
+        response = new_session.get(GLOBAL_IP_API_1)
+        ip_list = re.findall(IP_PATTERN, response.text)
         new_session.close()
         return ip_list[0]
     except Exception as e:
-        if DEBUG:
+        if debug:
             print(str(e))
-        return "Error"
+        return GENERAL_ERROR_MESSAGE
 
 
-def set_ip(ip, restart=False, device="eth0", DEBUG=False):
+def set_ip(ip, restart=False, device="eth0", debug=False):
     """
     Set static ip in interfaces file (need sudo).
 
+    :param ip: static ip
+    :type ip :str
     :param restart : restart flag
     :type restart : bool
     :param device: network device name
     :type device:str
-    :param ip: static ip
-    :type ip :str
-    :param DEBUG: flag for using Debug mode
-    :type DEBUG:bool
+    :param debug: flag for using debug mode
+    :type debug:bool
     :return: True in successful
     """
     static_string = '''
@@ -111,7 +110,7 @@ iface device inet static
         dns-nameservers 8.8.8.8 8.8.4.4
     '''
     try:
-        if bool(re.match(ip_pattern, ip)) == False or ip.find(
+        if not bool(re.match(IP_PATTERN, ip)) or ip.find(
                 "192.168.") == -1 or device not in mac().keys():
             raise Exception("IP Formation Error")
         static_string = static_string.replace("ip", ip)
@@ -125,25 +124,25 @@ iface device inet static
             restart_func()
         return True
     except Exception as e:
-        if DEBUG:
+        if debug:
             print(str(e))
-        return "Error"
+        return GENERAL_ERROR_MESSAGE
 
 
-def ping(ip, packet_number=3, DEBUG=False):
+def ping(ip, packet_number=3, debug=False):
     """
     Ping ip and return True if this ip is available and False otherwise.
 
     :param ip: target ip
-    :param packet_number: number of packet to size
-    :param DEBUG: flag for using Debug mode
     :type ip :str
+    :param packet_number: number of packet to size
     :type packet_number:int
-    :type DEBUG:bool
+    :param debug: flag for using debug mode
+    :type debug:bool
     :return: a boolean value (True if ip is available and False otherwise)
     """
     try:
-        if re.match(ip_pattern, ip) == False:
+        if not re.match(IP_PATTERN, ip):
             raise Exception("IP Formation Error")
         output = str(list(sub.Popen(["ping",
                                      ip,
@@ -155,17 +154,17 @@ def ping(ip, packet_number=3, DEBUG=False):
             return True
         return False
     except Exception as e:
-        if DEBUG:
+        if debug:
             print(str(e))
-        return "Error"
+        return GENERAL_ERROR_MESSAGE
 
 
-def mac(DEBUG=False):
+def mac(debug=False):
     """
     Return mac addresses of net devices.
 
-    :param DEBUG: flag for using Debug mode
-    :type DEBUG:bool
+    :param debug: flag for using debug mode
+    :type debug:bool
     :return: return mac addresses as dict with name as keys and mac addresses as values
     """
     try:
@@ -178,6 +177,6 @@ def mac(DEBUG=False):
             mac_addr.close()
         return dict(zip(dir_list, mac_list))
     except Exception as e:
-        if DEBUG:
+        if debug:
             print(str(e))
-        return "Error"
+        return GENERAL_ERROR_MESSAGE