+++ orangetool/orangetool_params.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+"""Orangetool params."""
+
+ORANGETOOL_VERSION = "0.45"
+UPDATE_URL = "http://www.orangetool.ir/version"
+IP_PATTERN = r"(?:[0-9]{1,3}\.){3}[0-9]{1,3}"
+GLOBAL_IP_API_1 = "http://ipinfo.io/ip"
+GENERAL_ERROR_MESSAGE = "Error"
+ROOT_ERROR_MESSAGE = "Root Error"