+++ orangetool/orangetool_storage.py
@@ -3,15 +3,18 @@
 import subprocess as sub
 import os
 import string
-import random
+from .orangetool_params import GENERAL_ERROR_MESSAGE, ROOT_ERROR_MESSAGE
+from .orangetool_utils import random_generator
 
 
-def mount_status(device_name, DEBUG=False):
+def mount_status(device_name, debug=False):
     """
     Return addresses of mounted memory devices in dev by device name.
 
-    :param DEBUG: flag for using Debug mode
-    :type DEBUG:bool
+    :param device_name: name of device
+    :type device_name: str
+    :param debug: flag for using debug mode
+    :type debug:bool
     :return: list of memory devices
     """
     try:
@@ -24,20 +27,19 @@ def mount_status(device_name, DEBUG=False):
                 memory_list.append(temp[1])
         if len(memory_list) == 0:
             return "u"
-        else:
-            return memory_list
+        return memory_list
     except Exception as e:
-        if DEBUG:
+        if debug:
             print(str(e))
-        return "Error"
+        return GENERAL_ERROR_MESSAGE
 
 
-def storage_status(DEBUG=False):
+def storage_status(debug=False):
     """
     Return all of the inserted memory and their status.
 
-    :param DEBUG: flag for using Debug mode
-    :type DEBUG:bool
+    :param debug: flag for using debug mode
+    :type debug:bool
     :return: all of the inserted memory and their status as dictionary ( device name as keys and mount status (mounted_addresses as list and u --> unmounted) as values
     """
     try:
@@ -51,40 +53,40 @@ def storage_status(DEBUG=False):
             memory_status.append(mount_status(item))
         return dict(zip(memory_items, memory_status))
     except Exception as e:
-        if DEBUG:
+        if debug:
             print(str(e))
-        return "Error"
+        return GENERAL_ERROR_MESSAGE
 
 
-def unmount(ADDRESS, DEBUG=False):
+def unmount(address, debug=False):
     """
     Unmount memory devices by addresses.
 
-    :param ADDRESS: address of that device mount on
-    :type ADDRESS:str
-    :param DEBUG: flag for using Debug mode
-    :type DEBUG:bool
+    :param address: address of that device mount on
+    :type address:str
+    :param debug: flag for using debug mode
+    :type debug:bool
     :return: True if device unmount correctly and False other wise
     """
     try:
-        command = sub.Popen(["umount", ADDRESS],
+        command = sub.Popen(["umount", address],
                             stdout=sub.PIPE, stderr=sub.PIPE)
         output = list(command.communicate())
         if len(output[0]) == 0 and len(output[1]) == 0:
             return True
         return False
     except Exception as e:
-        if DEBUG:
+        if debug:
             print(str(e))
-        return "Error"
+        return GENERAL_ERROR_MESSAGE
 
 
-def unmount_all(DEBUG=False):
+def unmount_all(debug=False):
     """
     Unmount all of the mounted devices.
 
-    :param DEBUG: flag for using Debug mode
-    :type DEBUG:bool
+    :param debug: flag for using debug mode
+    :type debug:bool
     :return: return True if all of the mounted devices unmount correctly
     """
     try:
@@ -98,37 +100,20 @@ def unmount_all(DEBUG=False):
                     unmount(j)
         return True
     except Exception as e:
-        if DEBUG:
+        if debug:
             print(str(e))
-        return "Error"
+        return GENERAL_ERROR_MESSAGE
 
-
-def random_generator(number):
-    """
-    Generate random number.
-
-    :param number: random number digits
-    :type number: int
-    :return: random number as str
-    """
-    response = ""
-    i = 0
-    while(i < number):
-        i += 1
-        response += str(random.randint(0, 9))
-    return response
-
-
-def mount(device_name, mount_address=None, DEBUG=False):
+def mount(device_name, mount_address=None, debug=False):
     """
     Mount memory devices by addresses.
 
     :param device_name: name of device for mounted example = sda1
-    :param mount_address: address for mounting device example = /mnt/usb , default value is None in this case function generate random number for mount folder name
-    :param DEBUG: flag for using Debug mode
     :type device_name:str
+    :param mount_address: address for mounting device example = /mnt/usb , default value is None in this case function generate random number for mount folder name
     :type mount_address:str
-    :type DEBUG:bool
+    :param debug: flag for using debug mode
+    :type debug:bool
     :return: True if device mount correctly and False other wise
     """
     try:
@@ -148,6 +133,53 @@ def mount(device_name, mount_address=None, DEBUG=False):
             return True
         return False
     except Exception as e:
-        if DEBUG:
+        if debug:
             print(str(e))
-        return "Error"
+        return GENERAL_ERROR_MESSAGE
+
+
+def usb_control(code, debug=False):
+    """
+    Control different usb options.
+
+    :param code: permission code
+    :type code: str
+    :param debug: flag for using debug mode
+    :type debug: bool
+    :return: None
+    """
+    try:
+        command = sub.Popen(
+            ["chmod", "-R", code, "/media/"],
+            stderr=sub.PIPE,
+            stdout=sub.PIPE,
+            stdin=sub.PIPE)
+        response = list(command.communicate())
+        if len(response[1]) > 0:
+            raise Exception(ROOT_ERROR_MESSAGE)
+    except Exception as e:
+        if debug:
+            print(str(e))
+        return GENERAL_ERROR_MESSAGE
+
+
+def usb_on(debug=False):
+    """
+    Shortcut for enable usb (need sudo).
+
+    :param debug: flag for using debug mode
+    :type debug:bool
+    :return: None
+    """
+    usb_control("777", debug)
+
+
+def usb_off(debug=False):
+    """
+    Shortcut for disable usb (need sudo).
+
+    :param debug: flag for using debug mode
+    :type debug:bool
+    :return: None
+    """
+    usb_control("000", debug)