+++ orangetool/orangetool_system.py
@@ -1,128 +1,101 @@
 # -*- coding: utf-8 -*-
 """Orangetool system functions."""
 import subprocess as sub
+from .orangetool_params import ORANGETOOL_VERSION, UPDATE_URL, GENERAL_ERROR_MESSAGE, ROOT_ERROR_MESSAGE
+from .orangetool_utils import time_convert
 import time
 import requests
 from art import tprint
-ip_pattern = r"(?:[0-9]{1,3}\.){3}[0-9]{1,3}"
-api_1 = "http://ipinfo.io/ip"
-VERSION = "0.35"
-UPDATE_URL = "http://www.orangetool.ir/version"
 
 
-def check_update(DEBUG=False):
+def check_update(debug=False):
     """
     Check orangetool site for new version.
 
-    :param DEBUG: flag for using Debug mode
-    :type DEBUG:bool
+    :param debug: flag for using debug mode
+    :type debug:bool
     :return: True if new version is available
     """
     try:
         new_version = requests.get(UPDATE_URL).text
-        if float(new_version) > float(VERSION):
+        if float(new_version) > float(ORANGETOOL_VERSION):
             print("New Version (" + new_version + ") Of Orangetool Is Available")
             return True
         print("Update!")
         return False
     except Exception as e:
-        if DEBUG:
+        if debug:
             print(str(e))
-        return "Error"
+        return GENERAL_ERROR_MESSAGE
 
 
-def get_temp(Zone=0, DEBUG=False):
+def get_temp(zone=0, debug=False):
     """
     Read cpu temperature.
 
-    :param DEBUG : flag for using Debug mode
-    :param Zone : thermal Zone Index
-    :type DEBUG:bool
-    :type Zone:int
+    :param zone : thermal zone index
+    :type zone:int
+    :param debug : flag for using debug mode
+    :type debug:bool
     :return: board temp as string in celsius
     """
     try:
-        command = open("/sys/class/thermal/thermal_zone" + str(Zone) + "/temp")
-        # command=sub.Popen(["cat","/sys/class/thermal/thermal_zone"+str(Zone)+"/temp"],stderr=sub.PIPE,stdin=sub.PIPE,stdout=sub.PIPE)
+        command = open("/sys/class/thermal/thermal_zone" + str(zone) + "/temp")
+        # command=sub.Popen(["cat","/sys/class/thermal/thermal_zone"+str(zone)+"/temp"],stderr=sub.PIPE,stdin=sub.PIPE,stdout=sub.PIPE)
         # response=list(command.communicate())
         response = command.read()
         return response[:-1]
         # if len(response[0])!=0:
         # return str(response[0])[2:-3]
         # else:
-        # return "Error"
+        # return GENERAL_ERROR_MESSAGE
     except Exception as e:
-        if DEBUG:
+        if debug:
             print(str(e))
-        return "Error"
+        return GENERAL_ERROR_MESSAGE
 
-
-def zero_insert(input_string):
+def time_control(mode="uptime",debug=False):
     """
-    Get a string as input if input is one digit add a zero.
+    Return system time.
 
-    :param input_string: input digit az string
-    :type input_string:str
-    :return: modified output as str
+    :param mode: time mode (uptime / idle)
+    :type mode: str
+    :param debug: flag for using debug mode
+    :type debug:bool
+    :return: system time as string
     """
-    if len(input_string) == 1:
-        return "0" + input_string
-    return input_string
+    index = 0
+    if mode  in ["idle","idletime"]:
+        index = 1
+    try:
+        command = open("/proc/uptime")
+        response = command.read()
+        return time_convert(response[:-1].split(" ")[index])
+    except Exception as e:
+        if debug:
+            print(str(e))
+        return GENERAL_ERROR_MESSAGE
 
-
-def time_convert(input_string):
-    """
-    Convert input_string from uptime from sec to DD,HH,MM,SS Format.
-
-    :param input_string: input time string  in sec
-    :type input_string:str
-    :return: converted time as string
-    """
-    input_sec = float(input_string)
-    input_minute = input_sec // 60
-    input_sec = int(input_sec - input_minute * 60)
-    input_hour = input_minute // 60
-    input_minute = int(input_minute - input_hour * 60)
-    input_day = int(input_hour // 24)
-    input_hour = int(input_hour - input_day * 24)
-    return zero_insert(str(input_day)) + " days, " + zero_insert(str(input_hour)) + " hour, " + \
-        zero_insert(str(input_minute)) + " minutes, " + zero_insert(str(input_sec)) + " seconds"
-
-
-def uptime(DEBUG=False):
+def uptime(debug=False):
     """
     Return system uptime.
 
-    :param DEBUG: flag for using Debug mode
-    :type DEBUG:bool
+    :param debug: flag for using debug mode
+    :type debug:bool
     :return: system uptime as string
     """
-    try:
-        command = open("/proc/uptime")
-        response = command.read()
-        return time_convert(response[:-1].split(" ")[0])
-    except Exception as e:
-        if DEBUG:
-            print(str(e))
-        return "Error"
+    return time_control(mode="uptime",debug=debug)
 
 
-def idletime(DEBUG=False):
+def idletime(debug=False):
     """
     Return system idletime.
 
-    :param DEBUG: flag for using Debug mode
-    :type DEBUG:bool
+    :param debug: flag for using debug mode
+    :type debug:bool
     :return: system idle as string
     """
-    try:
-        command = open("/proc/uptime")
-        response = command.read()
-        return time_convert(response[:-1].split(" ")[1])
-    except Exception as e:
-        if DEBUG:
-            print(str(e))
-        return "Error"
+    return time_control(mode="idle", debug=debug)
 
 
 def version():
@@ -132,21 +105,21 @@ def version():
     :return: return orangetool-version number as string
     """
     tprint("orangetool", font="bulbhead")
-    tprint("v"+VERSION,font="bulbhead")
+    tprint("v" + ORANGETOOL_VERSION, font="bulbhead")
 
 
-def wakeup(day=0, hour=0, minute=0, DEBUG=False):
+def wakeup(day=0, hour=0, minute=0, debug=False):
     """
     Set wakeup time for kernel RTC (need sudo).
 
     :param day: days for wakeup
-    :param hour: hout for wakeup
-    :param minute: minute for wakeup
-    :param DEBUG: flag for using Debug mode
     :type day:int
+    :param hour: hour for wakeup
     :type hour:int
+    :param minute: minute for wakeup
     :type minute:int
-    :type DEBUG:bool
+    :param debug: flag for using debug mode
+    :type debug:bool
     :return: bool
     """
     try:
@@ -160,19 +133,19 @@ def wakeup(day=0, hour=0, minute=0, DEBUG=False):
         file.close()
         return True
     except Exception as e:
-        if DEBUG:
+        if debug:
             print(str(e))
-        return "Error"
+        return GENERAL_ERROR_MESSAGE
 
 
-def power_control(command, DEBUG=False):
+def power_control(command, debug=False):
     """
     Control different power options.
 
     :param command: input command
     :type command: str
-    :param DEBUG: flag for using Debug mode
-    :type DEBUG: bool
+    :param debug: flag for using debug mode
+    :type debug: bool
     :return: None
     """
     try:
@@ -183,41 +156,52 @@ def power_control(command, DEBUG=False):
             stdin=sub.PIPE)
         response = list(command.communicate())
         if len(response[1]) > 0:
-            raise Exception('Root Error')
+            raise Exception(ROOT_ERROR_MESSAGE)
     except Exception as e:
-        if DEBUG:
+        if debug:
             print(str(e))
-        return "Error"
+        return GENERAL_ERROR_MESSAGE
 
 
-def sleep(DEBUG=False):
+def sleep(debug=False):
     """
     Shortcut for sleep command (need sudo).
 
-    :param DEBUG: flag for using Debug mode
-    :type DEBUG:bool
+    :param debug: flag for using debug mode
+    :type debug:bool
     :return: None
     """
-    power_control("pm-suspend", DEBUG)
+    power_control("pm-suspend", debug)
 
 
-def halt(DEBUG=False):
+def hibernate(debug=False):
+    """
+    Shortcut for hibernate command (need sudo).
+
+    :param debug: flag for using debug mode
+    :type debug:bool
+    :return: None
+    """
+    power_control("pm-hibernate", debug)
+
+
+def halt(debug=False):
     """
     Shortcut for poweroff (need sudo).
 
-    :param DEBUG: flag for using Debug mode
-    :type DEBUG:bool
+    :param debug: flag for using debug mode
+    :type debug:bool
     :return: None
     """
-    power_control("poweroff", DEBUG)
+    power_control("poweroff", debug)
 
 
-def restart(DEBUG=False):
+def restart(debug=False):
     """
     Shortcut for reboot (need sudo).
 
-    :param DEBUG: flag for using Debug mode
-    :type DEBUG:bool
+    :param debug: flag for using debug mode
+    :type debug:bool
     :return: None
     """
-    power_control("reboot", DEBUG)
+    power_control("reboot", debug)