+++ orangetool/orangetool_test.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+"""Orangetool test."""
+from orangetool import version, check_update
+
+if __name__ == "__main__":
+    version()
+    check_update()